Ecommerce V1.0
This project involves creating the frontend of an e-commerce website using HTML, CSS, JavaScript, jQuery, and Bootstrap. The goal is to develop a visually appealing and user-friendly interface that allows users to browse, search, and purchase products. The website will include various pages such as the homepage, product listing, product details, shopping cart, and checkout.